+++ Example/ViewDeckExample/ViewController.m
@@ -181,7 +181,7 @@
     }
     
     cell.textLabel.textAlignment = indexPath.section ? UITextAlignmentRight : UITextAlignmentLeft;
-    cell.textLabel.text = [NSString stringWithFormat:@"ledge: %d", indexPath.row*44];
+    cell.textLabel.text = [NSString stringWithFormat:@"ledge: %d", (int)indexPath.row*44];
     
     return cell;
 }
